refactor: split schema operator and roundtrip roots
Break annotation, relational inference, operator workflow, and\nroundtrip provider orchestration into narrower helper families while\nkeeping the public roots stable.\n\n🤖 Generated with [Claude Code](https://claude.com/claude-code)\n\nCo-Authored-By: Claude <noreply@anthropic.com>
