+++ polylogue/schemas/generation_dynamic_keys.py
@@ -1,0 +1,98 @@
+"""Dynamic-key collapse helpers for schema generation."""
+
+from __future__ import annotations
+
+from typing import Any
+
+try:
+    from genson import SchemaBuilder
+
+    GENSON_AVAILABLE = True
+except ImportError:
+    SchemaBuilder = None
+    GENSON_AVAILABLE = False
+
+from polylogue.schemas.field_stats import is_dynamic_key
+
+_HIGH_CARDINALITY_KEY_THRESHOLD = 128
+_PATHLIKE_KEY_RATIO_THRESHOLD = 0.35
+
+
+def _looks_pathlike_key(key: str) -> bool:
+    if "/" in key or "\\" in key:
+        return True
+    if key.count(".") >= 2:
+        return True
+    return ":" in key and len(key) > 2
+
+
+def _should_collapse_high_cardinality_keys(keys: list[str]) -> bool:
+    if len(keys) >= _HIGH_CARDINALITY_KEY_THRESHOLD:
+        return True
+    if len(keys) < 24:
+        return False
+    pathlike = sum(1 for key in keys if _looks_pathlike_key(key))
+    return (pathlike / len(keys)) >= _PATHLIKE_KEY_RATIO_THRESHOLD
+
+
+def merge_schemas(schemas: list[dict[str, Any]]) -> dict[str, Any]:
+    """Merge multiple schemas into one using genson when available."""
+    if not GENSON_AVAILABLE:
+        return schemas[0] if schemas else {}
+    builder = SchemaBuilder()
+    for schema in schemas:
+        builder.add_schema(schema)
+    return dict(builder.to_schema())
+
+
+def collapse_dynamic_keys(schema: dict[str, Any]) -> dict[str, Any]:
+    """Collapse dynamic key properties into additionalProperties."""
+    if not isinstance(schema, dict):
+        return schema
+
+    if "properties" in schema:
+        props = schema["properties"]
+        static_props: dict[str, Any] = {}
+        dynamic_schemas: list[dict[str, Any]] = []
+        key_names = list(props.keys())
+
+        for key, value in props.items():
+            collapsed_value = collapse_dynamic_keys(value)
+            if is_dynamic_key(key):
+                dynamic_schemas.append(collapsed_value)
+            else:
+                static_props[key] = collapsed_value
+
+        if _should_collapse_high_cardinality_keys(key_names):
+            dynamic_schemas.extend(static_props.values())
+            static_props = {}
+            schema["x-polylogue-high-cardinality-keys"] = True
+
+        if dynamic_schemas:
+            schema["properties"] = static_props
+            schema["additionalProperties"] = merge_schemas(dynamic_schemas)
+            schema["x-polylogue-dynamic-keys"] = True
+            if "required" in schema:
+                schema["required"] = [required for required in schema["required"] if required in static_props]
+        else:
+            schema["properties"] = static_props
+
+    if "items" in schema:
+        schema["items"] = collapse_dynamic_keys(schema["items"])
+
+    for keyword in ("anyOf", "oneOf", "allOf"):
+        if keyword in schema:
+            schema[keyword] = [collapse_dynamic_keys(item) for item in schema[keyword]]
+
+    if "additionalProperties" in schema and isinstance(schema["additionalProperties"], dict):
+        schema["additionalProperties"] = collapse_dynamic_keys(schema["additionalProperties"])
+
+    return schema
+
+
+__all__ = [
+    "GENSON_AVAILABLE",
+    "SchemaBuilder",
+    "collapse_dynamic_keys",
+    "merge_schemas",
+]
